+++ spec/webauthn/authenticator_attestation_response_spec.rb
@@ -3,7 +3,6 @@
 require "spec_helper"
 require "support/seeds"
 
-require "base64"
 require "webauthn/authenticator_attestation_response"
 require "openssl"
 
@@ -24,11 +23,7 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
 
   let(:public_key_credential) { client.create(challenge: original_challenge) }
 
-  before do
-    WebAuthn.configuration.origin = origin
-  end
-
-  context "when everything's in place" do
+  shared_examples "a valid attestation response" do
     it "verifies" do
       expect(attestation_response.verify(original_challenge)).to be_truthy
     end
@@ -36,62 +31,161 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
     it "is valid" do
       expect(attestation_response.valid?(original_challenge)).to be_truthy
     end
+  end
 
-    # TODO: let FakeClient#create recieve a fixed credential
-    # https://github.com/cedarcode/webauthn-ruby/pull/302#discussion_r365338434
-    it "returns the credential" do
-      credential = attestation_response.credential
+  context "when everything's in place" do
+    context "when there is a single origin" do
+      before do
+        WebAuthn.configuration.allowed_origins = [origin]
+      end
 
-      expect(credential.id.class).to eq(BinData::String)
-      expect(credential.id.encoding).to eq(Encoding::BINARY)
-      expect(credential.public_key.class).to eq(String)
-      expect(credential.public_key.encoding).to be(Encoding::BINARY)
+      it_behaves_like "a valid attestation response"
+
+      # TODO: let FakeClient#create recieve a fixed credential
+      # https://github.com/cedarcode/webauthn-ruby/pull/302#discussion_r365338434
+      it "returns the credential" do
+        credential = attestation_response.credential
+
+        expect(credential.id.class).to eq(BinData::String)
+        expect(credential.id.encoding).to eq(Encoding::BINARY)
+        expect(credential.public_key.class).to eq(String)
+        expect(credential.public_key.encoding).to be(Encoding::BINARY)
+      end
+    end
+
+    context "when there are multiple allowed origins" do
+      let(:allowed_origins) do
+        [
+          fake_origin,
+          "android:apk-key-hash:blablablablablalblalla"
+        ]
+      end
+
+      before do
+        WebAuthn.configuration.allowed_origins = allowed_origins
+      end
+
+      context "when rp_id is set explicitly" do
+        before do
+          WebAuthn.configuration.rp_id = "localhost"
+        end
+
+        it_behaves_like "a valid attestation response"
+
+        # TODO: let FakeClient#create recieve a fixed credential
+        # https://github.com/cedarcode/webauthn-ruby/pull/302#discussion_r365338434
+        it "returns the credential" do
+          credential = attestation_response.credential
+
+          expect(credential.id.class).to eq(BinData::String)
+          expect(credential.id.encoding).to eq(Encoding::BINARY)
+          expect(credential.public_key.class).to eq(String)
+          expect(credential.public_key.encoding).to be(Encoding::BINARY)
+        end
+      end
+
+      context "when rp_id is not set explicitly" do
+        before do
+          WebAuthn.configuration.rp_id = nil
+        end
+
+        it "raises error" do
+          expect { attestation_response.verify(original_challenge) }.to raise_error(WebAuthn::RpIdVerificationError)
+        end
+
+        it "is not valid" do
+          expect(attestation_response.valid?(original_challenge)).to be_falsey
+        end
+
+        # TODO: let FakeClient#create recieve a fixed credential
+        # https://github.com/cedarcode/webauthn-ruby/pull/302#discussion_r365338434
+        it "returns the credential" do
+          credential = attestation_response.credential
+
+          expect(credential.id.class).to eq(BinData::String)
+          expect(credential.id.encoding).to eq(Encoding::BINARY)
+          expect(credential.public_key.class).to eq(String)
+          expect(credential.public_key.encoding).to be(Encoding::BINARY)
+        end
+      end
     end
   end
 
   context "when fido-u2f attestation" do
     let(:original_challenge) do
-      Base64.strict_decode64(seeds[:security_key_direct][:credential_creation_options][:challenge])
+      WebAuthn::Encoders::Base64Encoder.decode(seeds[:security_key_direct][:credential_creation_options][:challenge])
     end
 
-    let(:origin) { "http://localhost:3000" }
+    context "when there is a single origin" do
+      let(:origin) { "http://localhost:3000" }
 
-    let(:attestation_response) do
-      response = seeds[:security_key_direct][:authenticator_attestation_response]
+      let(:attestation_response) do
+        response = seeds[:security_key_direct][:authenticator_attestation_response]
 
-      WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.strict_decode64(response[:attestation_object]),
-        client_data_json: Base64.strict_decode64(response[:client_data_json])
-      )
+        WebAuthn::AuthenticatorAttestationResponse.new(
+          attestation_object: WebAuthn::Encoders::Base64Encoder.decode(response[:attestation_object]),
+          client_data_json: WebAuthn::Encoders::Base64Encoder.decode(response[:client_data_json])
+        )
+      end
+
+      before do
+        WebAuthn.configuration.attestation_root_certificates_finders = finder_for('feitian_ft_fido_0200.pem')
+        WebAuthn.configuration.allowed_origins = [origin]
+      end
+
+      it_behaves_like "a valid attestation response"
+
+      it "returns attestation info" do
+        attestation_response.valid?(original_challenge)
+
+        expect(attestation_response.attestation_type).to eq("Basic_or_AttCA")
+        expect(attestation_response.attestation_trust_path).to all(be_kind_of(OpenSSL::X509::Certificate))
+      end
+
+      it "returns the credential" do
+        expect(attestation_response.credential.id.length).to be >= 16
+      end
+
+      it "returns the attestation certificate key" do
+        expect(attestation_response.attestation_certificate_key_id).to(
+          eq("f4b64a68c334e901b8e23c6e66e6866c31931f5d")
+        )
+      end
     end
 
-    before do
-      WebAuthn.configuration.attestation_root_certificates_finders = finder_for('feitian_ft_fido_0200.pem')
-    end
+    context "when there are multiple allowed origins" do
+      let(:allowed_origins) do
+        [
+          fake_origin,
+          "android:apk-key-hash:blablablablablalblalla"
+        ]
+      end
 
-    it "verifies" do
-      expect(attestation_response.verify(original_challenge)).to be_truthy
-    end
+      before do
+        WebAuthn.configuration.allowed_origins = allowed_origins
+      end
 
-    it "is valid" do
-      expect(attestation_response.valid?(original_challenge)).to eq(true)
-    end
+      context "when rp_id is set explicitly" do
+        before do
+          WebAuthn.configuration.rp_id = "localhost"
+        end
 
-    it "returns attestation info" do
-      attestation_response.valid?(original_challenge)
+        it_behaves_like "a valid attestation response"
+      end
 
-      expect(attestation_response.attestation_type).to eq("Basic_or_AttCA")
-      expect(attestation_response.attestation_trust_path).to all(be_kind_of(OpenSSL::X509::Certificate))
-    end
+      context "when rp_id is not set explicitly" do
+        before do
+          WebAuthn.configuration.rp_id = nil
+        end
 
-    it "returns the credential" do
-      expect(attestation_response.credential.id.length).to be >= 16
-    end
+        it "raises error" do
+          expect { attestation_response.verify(original_challenge) }.to raise_error(WebAuthn::RpIdVerificationError)
+        end
 
-    it "returns the attestation certificate key" do
-      expect(attestation_response.attestation_certificate_key_id).to(
-        eq("f4b64a68c334e901b8e23c6e66e6866c31931f5d")
-      )
+        it "is not valid" do
+          expect(attestation_response.valid?(original_challenge)).to be_falsey
+        end
+      end
     end
   end
 
@@ -99,7 +193,7 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
     let(:origin) { "https://localhost:13010" }
 
     let(:original_challenge) do
-      Base64.strict_decode64(
+      WebAuthn::Encoders::Base64Encoder.decode(
         seeds[:security_key_packed_self][:credential_creation_options][:challenge]
       )
     end
@@ -108,18 +202,16 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
       response = seeds[:security_key_packed_self][:authenticator_attestation_response]
 
       WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.strict_decode64(response[:attestation_object]),
-        client_data_json: Base64.strict_decode64(response[:client_data_json])
+        attestation_object: WebAuthn::Encoders::Base64Encoder.decode(response[:attestation_object]),
+        client_data_json: WebAuthn::Encoders::Base64Encoder.decode(response[:client_data_json])
       )
     end
 
-    it "verifies" do
-      expect(attestation_response.verify(original_challenge)).to be_truthy
+    before do
+      WebAuthn.configuration.allowed_origins = [origin]
     end
 
-    it "is valid" do
-      expect(attestation_response.valid?(original_challenge)).to eq(true)
-    end
+    it_behaves_like "a valid attestation response"
 
     it "returns attestation info" do
       attestation_response.valid?(original_challenge)
@@ -141,7 +233,7 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
     let(:origin) { "http://localhost:3000" }
 
     let(:original_challenge) do
-      Base64.strict_decode64(
+      WebAuthn::Encoders::Base64Encoder.decode(
         seeds[:security_key_packed_x5c][:credential_creation_options][:challenge]
       )
     end
@@ -150,22 +242,17 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
       response = seeds[:security_key_packed_x5c][:authenticator_attestation_response]
 
       WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.strict_decode64(response[:attestation_object]),
-        client_data_json: Base64.strict_decode64(response[:client_data_json])
+        attestation_object: WebAuthn::Encoders::Base64Encoder.decode(response[:attestation_object]),
+        client_data_json: WebAuthn::Encoders::Base64Encoder.decode(response[:client_data_json])
       )
     end
 
     before do
       WebAuthn.configuration.attestation_root_certificates_finders = finder_for('yubico_u2f_root.pem')
+      WebAuthn.configuration.allowed_origins = [origin]
     end
 
-    it "verifies" do
-      expect(attestation_response.verify(original_challenge)).to be_truthy
-    end
-
-    it "is valid" do
-      expect(attestation_response.valid?(original_challenge)).to eq(true)
-    end
+    it_behaves_like "a valid attestation response"
 
     it "returns attestation info" do
       attestation_response.valid?(original_challenge)
@@ -186,18 +273,21 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
   context "when TPM attestation" do
     let(:origin) { seeds[:tpm][:origin] }
     let(:time) { Time.utc(2019, 8, 13, 22, 6) }
-    let(:challenge) { Base64.strict_decode64(seeds[:tpm][:credential_creation_options][:challenge]) }
+    let(:challenge) do
+      WebAuthn::Encoders::Base64UrlEncoder.decode(seeds[:tpm][:credential_creation_options][:challenge])
+    end
 
     let(:attestation_response) do
       response = seeds[:tpm][:authenticator_attestation_response]
 
       WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.strict_decode64(response[:attestation_object]),
-        client_data_json: Base64.strict_decode64(response[:client_data_json])
+        attestation_object: WebAuthn::Encoders::Base64UrlEncoder.decode(response[:attestation_object]),
+        client_data_json: WebAuthn::Encoders::Base64UrlEncoder.decode(response[:client_data_json])
       )
     end
 
     before do
+      WebAuthn.configuration.allowed_origins = [origin]
       WebAuthn.configure do |config|
         config.algorithms.concat(%w(RS1))
       end
@@ -216,15 +306,15 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
     end
 
     it "verifies" do
-      expect(attestation_response.verify(challenge, origin)).to be_truthy
+      expect(attestation_response.verify(challenge, WebAuthn.configuration.allowed_origins)).to be_truthy
     end
 
     it "is valid" do
-      expect(attestation_response.valid?(challenge, origin)).to eq(true)
+      expect(attestation_response.valid?(challenge, WebAuthn.configuration.allowed_origins)).to eq(true)
     end
 
     it "returns attestation info" do
-      attestation_response.valid?(challenge, origin)
+      attestation_response.valid?(challenge, WebAuthn.configuration.allowed_origins)
 
       expect(attestation_response.attestation_type).to eq("AttCA")
       expect(attestation_response.attestation_trust_path).to all(be_kind_of(OpenSSL::X509::Certificate))
@@ -245,29 +335,26 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
     let(:origin) { "https://7f41ac45.ngrok.io" }
 
     let(:original_challenge) do
-      Base64.strict_decode64(seeds[:android_safetynet_direct][:credential_creation_options][:challenge])
+      WebAuthn::Encoders::Base64Encoder.decode(
+        seeds[:android_safetynet_direct][:credential_creation_options][:challenge]
+      )
     end
 
     let(:attestation_response) do
       response = seeds[:android_safetynet_direct][:authenticator_attestation_response]
 
       WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.strict_decode64(response[:attestation_object]),
-        client_data_json: Base64.strict_decode64(response[:client_data_json])
+        attestation_object: WebAuthn::Encoders::Base64Encoder.decode(response[:attestation_object]),
+        client_data_json: WebAuthn::Encoders::Base64Encoder.decode(response[:client_data_json])
       )
     end
 
     before do
+      WebAuthn.configuration.allowed_origins = [origin]
       allow(attestation_response.attestation_statement).to receive(:time).and_return(time)
     end
 
-    it "verifies" do
-      expect(attestation_response.verify(original_challenge)).to be_truthy
-    end
-
-    it "is valid" do
-      expect(attestation_response.valid?(original_challenge)).to eq(true)
-    end
+    it_behaves_like "a valid attestation response"
 
     it "returns attestation info" do
       attestation_response.valid?(original_challenge)
@@ -286,18 +373,16 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
   end
 
   context "when android-key attestation" do
-    let(:origin) { seeds[:android_key_direct][:origin] }
-
     let(:original_challenge) do
-      Base64.urlsafe_decode64(seeds[:android_key_direct][:credential_creation_options][:challenge])
+      WebAuthn::Encoders::Base64UrlEncoder.decode(seeds[:android_key_direct][:credential_creation_options][:challenge])
     end
 
     let(:attestation_response) do
       response = seeds[:android_key_direct][:authenticator_attestation_response]
 
       WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.urlsafe_decode64(response[:attestation_object]),
-        client_data_json: Base64.urlsafe_decode64(response[:client_data_json])
+        attestation_object: WebAuthn::Encoders::Base64UrlEncoder.decode(response[:attestation_object]),
+        client_data_json: WebAuthn::Encoders::Base64UrlEncoder.decode(response[:client_data_json])
       )
     end
 
@@ -305,27 +390,80 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
       WebAuthn.configuration.attestation_root_certificates_finders = finder_for('android_key_root.pem')
     end
 
-    it "verifies" do
-      expect(attestation_response.verify(original_challenge)).to be_truthy
+    context "when there is a single origin" do
+      let(:origin) { seeds[:android_key_direct][:origin] }
+
+      before do
+        WebAuthn.configuration.allowed_origins = [origin]
+      end
+
+      it_behaves_like "a valid attestation response"
+
+      it "returns attestation info" do
+        attestation_response.valid?(original_challenge)
+
+        expect(attestation_response.attestation_type).to eq("Basic")
+        expect(attestation_response.attestation_trust_path).to all(be_kind_of(OpenSSL::X509::Certificate))
+      end
+
+      it "returns the credential" do
+        expect(attestation_response.credential.id.length).to be >= 16
+      end
+
+      it "returns the AAGUID" do
+        expect(attestation_response.aaguid).to eq("550e4b54-aa47-409f-9a95-1ab76c130131")
+      end
     end
 
-    it "is valid" do
-      expect(attestation_response.valid?(original_challenge)).to eq(true)
-    end
+    context "when there are multiple allowed origins" do
+      let(:allowed_origins) do
+        [
+          seeds[:android_key_direct][:origin],
+          "android:apk-key-hash:blablablablablalblalla",
+          "localhost"
+        ]
+      end
 
-    it "returns attestation info" do
-      attestation_response.valid?(original_challenge)
+      before do
+        WebAuthn.configuration.allowed_origins = allowed_origins
+      end
 
-      expect(attestation_response.attestation_type).to eq("Basic")
-      expect(attestation_response.attestation_trust_path).to all(be_kind_of(OpenSSL::X509::Certificate))
-    end
+      context "when rp_id is set explicitly" do
+        before do
+          WebAuthn.configuration.rp_id = "localhost"
+        end
 
-    it "returns the credential" do
-      expect(attestation_response.credential.id.length).to be >= 16
-    end
+        it_behaves_like "a valid attestation response"
 
-    it "returns the AAGUID" do
-      expect(attestation_response.aaguid).to eq("550e4b54-aa47-409f-9a95-1ab76c130131")
+        it "returns attestation info" do
+          attestation_response.valid?(original_challenge)
+
+          expect(attestation_response.attestation_type).to eq("Basic")
+          expect(attestation_response.attestation_trust_path).to all(be_kind_of(OpenSSL::X509::Certificate))
+        end
+
+        it "returns the credential" do
+          expect(attestation_response.credential.id.length).to be >= 16
+        end
+
+        it "returns the AAGUID" do
+          expect(attestation_response.aaguid).to eq("550e4b54-aa47-409f-9a95-1ab76c130131")
+        end
+      end
+
+      context "when rp_id is not set explicitly" do
+        before do
+          WebAuthn.configuration.rp_id = nil
+        end
+
+        it "raises error" do
+          expect { attestation_response.verify(original_challenge) }.to raise_error(WebAuthn::RpIdVerificationError)
+        end
+
+        it "is not valid" do
+          expect(attestation_response.valid?(original_challenge)).to be_falsey
+        end
+      end
     end
   end
 
@@ -333,31 +471,27 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
     let(:origin) { seeds[:macbook_touch_id][:origin] }
 
     let(:original_challenge) do
-      Base64.urlsafe_decode64(seeds[:macbook_touch_id][:credential_creation_options][:challenge])
+      WebAuthn::Encoders::Base64UrlEncoder.decode(seeds[:macbook_touch_id][:credential_creation_options][:challenge])
     end
 
     let(:attestation_response) do
       response = seeds[:macbook_touch_id][:authenticator_attestation_response]
 
       WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.urlsafe_decode64(response[:attestation_object]),
-        client_data_json: Base64.urlsafe_decode64(response[:client_data_json])
+        attestation_object: WebAuthn::Encoders::Base64UrlEncoder.decode(response[:attestation_object]),
+        client_data_json: WebAuthn::Encoders::Base64UrlEncoder.decode(response[:client_data_json])
       )
     end
 
     before do
+      WebAuthn.configuration.allowed_origins = [origin]
+
       # Apple credential certificate expires after 3 days apparently.
       # Seed data was obtained 22nd Feb 2021, so we are simulating validation within that 3 day timeframe
       fake_certificate_chain_validation_time(attestation_response.attestation_statement, Time.parse("2021-02-23"))
     end
 
-    it "verifies" do
-      expect(attestation_response.verify(original_challenge)).to be_truthy
-    end
-
-    it "is valid" do
-      expect(attestation_response.valid?(original_challenge)).to eq(true)
-    end
+    it_behaves_like "a valid attestation response"
 
     it "returns attestation info" do
       attestation_response.valid?(original_challenge)
@@ -371,15 +505,21 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
     end
   end
 
-  it "returns user-friendly error if no client data received" do
-    attestation_response = WebAuthn::AuthenticatorAttestationResponse.new(
-      attestation_object: "",
-      client_data_json: nil
-    )
+  context "when no client data received" do
+    before do
+      WebAuthn.configuration.allowed_origins = [origin]
+    end
 
-    expect {
-      attestation_response.valid?("", "")
-    }.to raise_exception(WebAuthn::ClientDataMissingError)
+    it "returns user-friendly error if no client data received" do
+      attestation_response = WebAuthn::AuthenticatorAttestationResponse.new(
+        attestation_object: "",
+        client_data_json: nil
+      )
+
+      expect {
+        attestation_response.valid?("", "")
+      }.to raise_exception(WebAuthn::ClientDataMissingError)
+    end
   end
 
   describe "origin validation" do
@@ -396,16 +536,14 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
       )
     end
 
+    before do
+      WebAuthn.configuration.allowed_origins = [origin]
+    end
+
     context "matches the default one" do
       let(:actual_origin) { "http://localhost" }
 
-      it "verifies" do
-        expect(attestation_response.verify(original_challenge)).to be_truthy
-      end
-
-      it "is valid" do
-        expect(attestation_response.valid?(original_challenge)).to be_truthy
-      end
+      it_behaves_like "a valid attestation response"
     end
 
     context "doesn't match the default one" do
@@ -436,16 +574,14 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
       )
     end
 
+    before do
+      WebAuthn.configuration.allowed_origins = [origin]
+    end
+
     context "matches the default one" do
       let(:rp_id) { "localhost" }
 
-      it "verifies" do
-        expect(attestation_response.verify(original_challenge)).to be_truthy
-      end
-
-      it "is valid" do
-        expect(attestation_response.valid?(original_challenge)).to be_truthy
-      end
+      it_behaves_like "a valid attestation response"
     end
 
     context "doesn't match the default one" do
@@ -469,13 +605,7 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
         WebAuthn.configuration.rp_id = rp_id
       end
 
-      it "verifies" do
-        expect(attestation_response.verify(original_challenge)).to be_truthy
-      end
-
-      it "is valid" do
-        expect(attestation_response.valid?(original_challenge)).to be_truthy
-      end
+      it_behaves_like "a valid attestation response"
     end
   end
 
@@ -486,17 +616,17 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
       context "when silent_authentication is not set" do
         it "doesn't verify if user presence is not set" do
           expect {
-            attestation_response.verify(original_challenge, origin)
+            attestation_response.verify(original_challenge, [origin])
           }.to raise_exception(WebAuthn::UserPresenceVerificationError)
         end
 
         it "verifies if user presence is not required" do
-          expect(attestation_response.verify(original_challenge, origin, user_presence: false)).to be_truthy
+          expect(attestation_response.verify(original_challenge, [origin], user_presence: false)).to be_truthy
         end
 
         it "doesn't verify if user presence is required" do
           expect {
-            attestation_response.verify(original_challenge, origin, user_presence: true)
+            attestation_response.verify(original_challenge, [origin], user_presence: true)
           }.to raise_exception(WebAuthn::UserPresenceVerificationError)
         end
       end
@@ -513,17 +643,17 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
 
         it "doesn't verify if user presence is not set" do
           expect {
-            attestation_response.verify(original_challenge, origin)
+            attestation_response.verify(original_challenge, [origin])
           }.to raise_exception(WebAuthn::UserPresenceVerificationError)
         end
 
         it "verifies if user presence is not required" do
-          expect(attestation_response.verify(original_challenge, origin, user_presence: false)).to be_truthy
+          expect(attestation_response.verify(original_challenge, [origin], user_presence: false)).to be_truthy
         end
 
         it "doesn't verify if user presence is required" do
           expect {
-            attestation_response.verify(original_challenge, origin, user_presence: true)
+            attestation_response.verify(original_challenge, [origin], user_presence: true)
           }.to raise_exception(WebAuthn::UserPresenceVerificationError)
         end
       end
@@ -539,16 +669,16 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
         end
 
         it "verifies if user presence is not set" do
-          expect(attestation_response.verify(original_challenge, origin)).to be_truthy
+          expect(attestation_response.verify(original_challenge, [origin])).to be_truthy
         end
 
         it "verifies if user presence is not required" do
-          expect(attestation_response.verify(original_challenge, origin, user_presence: false)).to be_truthy
+          expect(attestation_response.verify(original_challenge, [origin], user_presence: false)).to be_truthy
         end
 
         it "doesn't verify if user presence is required" do
           expect {
-            attestation_response.verify(original_challenge, origin, user_presence: true)
+            attestation_response.verify(original_challenge, [origin], user_presence: true)
           }.to raise_exception(WebAuthn::UserPresenceVerificationError)
         end
       end
@@ -556,24 +686,36 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
   end
 
   describe "user verification" do
+    before do
+      WebAuthn.configuration.allowed_origins = [origin]
+    end
+
     context "when UV is not set" do
       let(:public_key_credential) { client.create(challenge: original_challenge, user_verified: false) }
 
       it "doesn't verify if user verification is required" do
         expect {
-          attestation_response.verify(original_challenge, origin, user_verification: true)
+          attestation_response.verify(
+            original_challenge,
+            WebAuthn.configuration.allowed_origins,
+            user_verification: true
+          )
         }.to raise_exception(WebAuthn::UserVerifiedVerificationError)
       end
     end
   end
 
   describe "attested credential data verification" do
+    before do
+      WebAuthn.configuration.allowed_origins = [origin]
+    end
+
     context "when AT is not set" do
       let(:public_key_credential) { client.create(challenge: original_challenge, attested_credential_data: false) }
 
       it "doesn't verify" do
         expect {
-          attestation_response.verify(original_challenge, origin)
+          attestation_response.verify(original_challenge, WebAuthn.configuration.allowed_origins)
         }.to raise_exception(WebAuthn::AttestedCredentialVerificationError)
       end
     end
@@ -585,7 +727,7 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
 
       it "doesn't verify" do
         expect {
-          attestation_response.verify(original_challenge, origin)
+          attestation_response.verify(original_challenge, WebAuthn.configuration.allowed_origins)
         }.to raise_exception(WebAuthn::AttestedCredentialVerificationError)
       end
     end
@@ -593,7 +735,7 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
 
   describe "attestation statement verification" do
     let(:original_challenge) do
-      Base64.strict_decode64(seeds[:security_key_direct][:credential_creation_options][:challenge])
+      WebAuthn::Encoders::Base64Encoder.decode(seeds[:security_key_direct][:credential_creation_options][:challenge])
     end
 
     let(:origin) { "http://localhost:3000" }
@@ -602,14 +744,15 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
       response = seeds[:security_key_direct][:authenticator_attestation_response]
 
       WebAuthn::AuthenticatorAttestationResponse.new(
-        attestation_object: Base64.strict_decode64(response[:attestation_object]),
-        client_data_json: Base64.strict_decode64(response[:client_data_json])
+        attestation_object: WebAuthn::Encoders::Base64Encoder.decode(response[:attestation_object]),
+        client_data_json: WebAuthn::Encoders::Base64Encoder.decode(response[:client_data_json])
       )
     end
 
     before do
       attestation_response.attestation_statement.instance_variable_get(:@statement)["sig"] =
         "corrupted signature".b
+      WebAuthn.configuration.allowed_origins = [origin]
     end
 
     context "when verification is set to true" do
@@ -617,7 +760,7 @@ RSpec.describe WebAuthn::AuthenticatorAttestationResponse do
         WebAuthn.configuration.verify_attestation_statement = true
       end
 
-      it "verifies the attestation statement" do
+      it "raises error" do
         expect { attestation_response.verify(original_challenge) }.to raise_error(OpenSSL::PKey::PKeyError)
       end
     end